Skull registration working. Tested with varying number of strays

+++ vtk_test/rigidregistration.h
@@ -2,6 +2,7 @@
 #define RIGIDREGISTRATION_H
 
 #include <Eigen/Dense>
+#include <QtGlobal>
 
 class RigidRegistration
 {
@@ -13,7 +14,7 @@ public:
     Eigen::MatrixXd GetTransform() const;
 
     // only available if pointRegisterOutliers() was used
-    Eigen::ArrayXd GetIndexMatch() const; // indices of correctly matched pts
+    Eigen::ArrayXi GetIndexMatch() const; // indices of correctly matched pts
     Eigen::MatrixXd getPtsMatch() const;    // actual matched pts
 
 private:
@@ -21,7 +22,7 @@ private:
     double          m_FRE;
 
     Eigen::MatrixXd m_ptsMatch; 
-    Eigen::ArrayXd m_indexMatch;
+    Eigen::ArrayXi m_indexMatch;
     
     friend void pointRegister(Eigen::MatrixXd ptsX, Eigen::MatrixXd ptsY, RigidRegistration &reg);
     friend void pointRegisterOutliers(Eigen::MatrixXd ptsX, Eigen::MatrixXd ptsY, RigidRegistration& reg);
@@ -29,8 +30,8 @@ private:
 
 void pointRegister(Eigen::MatrixXd ptsX, Eigen::MatrixXd ptsY, RigidRegistration& reg); // ordered point sets of the same size
 void pointRegisterOutliers(Eigen::MatrixXd ptsX, Eigen::MatrixXd ptsY, RigidRegistration& reg); // determines which points in ptsX correspond to those in ptsY
-int factorial(int n);
-int numPermutations(int n, int k);
-void createPermutations(int n, int k, Eigen::MatrixXd & iPerms);
+quint64 factorial(int n);
+quint32 numPermutations(int n, int k);
+void createPermutations(int n, int k, Eigen::MatrixXi & iPerms);
 
 #endif // RIGIDREGISTRATION_H
